Move question components into a dedicated component folder.
